Module neuro::data

Simple interfaces to import data.

Structs

ImageDataSet
Structure representing a collection of images.

ImageDataSetBuilder

ImageOps
Contains the parameters of the different operations applied on the images.

TabularDataSet
Structure representing tabular data.

Enums

DataSetError
Errors that may be raised by data sets methods.

Scaling
Defines the type of scaling that has been performed.

Traits

DataSet
Trait that must be implemented for any type of dataset supported by neuro.