Struct neuro::data::ImageDataSetBuilder

pub struct ImageDataSetBuilder { /* fields omitted */ }

Methods

impl ImageDataSetBuilder

pub fn from_dir(
path: &'static Path,
image_size: (u32, u32)
) -> ImageDataSetBuilder

Creates a dataset builder from a directory tree.

The images must be in folders named after the corresponding class in a train top-level directory. Optionally, if a test directory exists, its content will be used to create a test set. For instance:

pets/
  train/
    cats/
      img1.jpg
      img2.jpg
      ...
    dogs/
      img1.jpg
      img2.jpg
      ...
  test/
    cats/
      img1.jpg
      img2.jpg
      ...
    dogs/
      img1.jpg
      img2.jpg
      ...

The images are resized to the given size using nearest-neighbor interpolation. The aspect ratio is not conserved.

Example

let path = Path::new("dataset/pets");
let data = ImageDataSetBuilder::from_dir(&path, (32, 32))
    .one_hot_encode()
    .valid_split(0.2)
    .scale(1./255.)
    .rotate(10, 0.1)
    .build()?;

pub fn build(self) -> Result<ImageDataSet, Error>

Builds an ImageDataSet from the image dataset builder.

pub fn hflip(self, prob: f64) -> ImageDataSetBuilder

Flips the images horizontally with the given probability.

pub fn vflip(self, prob: f64) -> ImageDataSetBuilder

Flips the images vertically with the given probability.

pub fn one_hot_encode(self) -> ImageDataSetBuilder

One hot encodes the labels.

pub fn rotate(self, angle: i32, prob: f64) -> ImageDataSetBuilder

Rotates the images by an angle drawn from a uniform distribution with bounds ±angle (in degrees). A rotation is applied with the given probability.

pub fn valid_split(self, valid_frac: f64) -> ImageDataSetBuilder

Splits the data into a training and validation sets.

pub fn scale(self, factor: PrimitiveType) -> ImageDataSetBuilder

Scales the images by multiplying each pixel by the given factor.