Struct neuro::models::Network

pub struct Network<'a, D, O, L> where
    D: DataSet,
    O: Optimizer,
    L: Loss,  { /* fields omitted */ }

Structure representing a neural network.

Methods

impl<'a, D, O, L> Network<'a, D, O, L> where
D: DataSet,
O: Optimizer,
L: Loss,

pub fn new(
data: &'a D,
loss_function: L,
optimizer: O,
metrics: Option<Vec<Metrics>>,
regularizer: Option<Regularizer>
) -> Network<D, O, L>

Creates an empty neural network.

Arguments

data: dataset used to train the neural network

loss_function: loss function minimized in the optimization process

optimizer: algorithm used to optimize the parameters of the network

metrics: (optional) vector of Metrics used to evaluate the model

regularizer: (optional) method used to regularize the model

pub fn add(&mut self, layer: Box<dyn Layer>)

Adds a layer to the network.

Arguments

layer: layer to be added

pub fn fit(&mut self, batch_size: u64, epochs: u64, print_loss: Option<u64>)

Fits the neural network with the training data.

The training data are shuffled at the beginning of each epoch, before batches are created.

Arguments

batch_size: size of the mini-batches used for training

epochs: number of epochs to train for

print_loss: number of epochs between two computations and printing of the validation loss

pub fn predict(&self, input: &Tensor) -> Tensor

Computes the output of the network for the given input.

Arguments

input: tensor of inputs. Multiple samples can be evaluated at once by stacking them along the fourth dimension of the tensor.

Returns

Tensor of the predicted output

pub fn predict_class(&self, input: &Tensor) -> Vec<(String, PrimitiveType)>

Predicts the class for the input.

Arguments

input: tensor of inputs. Multiple samples can be evaluated at once by stacking them along the fourth dimension of the tensor.

Returns

Vector of tuples containing the predicted class and the probability for each sample.