Module neuro::tensor

Wrapper around ArrayFire's array with additional helper methods.

Enums

Reduction
Defines reduction methods.

Traits

TensorTrait
Defines additional methods for the Tensor type.

Type Definitions

Dim
Type alias for ArrayFire's Dim4.

PrimitiveType
This type is defined to easily change between f32 and f64 as the primitive type used by the crate. It has important consequences on the memory footprint of the crate when running deep and/or wide neural networks.

Tensor
Type alias for ArrayFire's Array.