Struct neuro::models::Network

pub struct Network { /* fields omitted */ }

Structure representing a neural network.

Methods

impl Network

pub fn new<L: 'static + Loss, O: 'static + Optimizer>(
input_shape: Dim,
loss_function: L,
optimizer: O,
regularizer: Option<Regularizer>
) -> Result<Network, Error>

Creates an empty neural network.

The input shape must be in the form [height, width, channel, 1]. Mini-batches are created along the fourth dimension.

pub fn add(&mut self, layer: Box<dyn Layer>)

Adds a layer to the network.

pub fn fit<T: DataSet>(
&mut self,
data: &T,
batch_size: u64,
epochs: u64,
print_loss: Option<u64>,
metrics: Option<Vec<Metrics>>
)

Fits the neural network with the training data.

The training data are shuffled at the beginning of each epoch, before batches are created. The progress is printed at every print_loss epoch.

pub fn evaluate<T: DataSet>(&self, data: &T, metrics: Option<Vec<Metrics>>)

Evaluates the model on the test set.

Arguments

data - The dataset containing the test data.

metrics - A vector containing the metrics that will be evaluated.

pub fn predict(&self, input: &Tensor) -> Tensor

Computes the output of the network for the given input.

Multiple samples can be evaluated at once by stacking them along the fourth dimension of the tensor.

Return value

Tensor of the predicted output

pub fn predict_class(&self, input: &Tensor) -> Vec<(String, PrimitiveType)>

Predicts the class for the input.

Multiple samples can be evaluated at once by stacking them along the fourth dimension of the tensor.

Return value

Vector of tuples containing the predicted class and the probability for each sample.

Panic

Panics if the model doesn't contain a classes dictionary.

pub fn save(&self, filename: &str) -> Result<(), Error>

Saves the model in HDF5 format.

pub fn load(filename: &str) -> Result<Network, Error>

Loads a model from a HDF5 file.